Move testcases to test_autostart.py
